backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
SQLAlchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.3
PyJWT>=2.9.0
python-multipart>=0.0.9
email-validator>=2.1.0
razorpay>=1.4.2
aiosqlite>=0.20.0
resend>=2.13.0
